+++ src/async-api/commons.ts
@@ -1,123 +1,174 @@
 import type {
   InfoObject as OASInfoObject,
-  ServerObject as OASServerObject,
   ServerVariableObject as OASServerVariableObject,
   ReferenceObject,
   SchemaObject,
 } from "openapi3-ts/oas31";
 import {
-  SocketIOChannelBinding,
-  SocketIOMessageBinding,
-  SocketIOOperationBinding,
-  SocketIOServerBinding,
-} from "./socket-io-binding";
+  WSChannelBinding,
+  WSMessageBinding,
+  WSOperationBinding,
+  WSServerBinding,
+} from "./ws-binding";
 
 /**
  * @fileoverview AsyncAPI specification
- * @version 2.6.0
+ * @version 3.0.0
  */
 
 interface Bindings<T> {
-  "socket.io"?: T;
+  ws?: T;
 }
 
-export interface ServerObject extends Omit<OASServerObject, "variables"> {
-  variables?: Record<string, ServerVariableObject>;
+/** @since 3.0.0 detached from OAS; added host, pathname, title, description, tags, externalDocs; changed security */
+export interface ServerObject {
+  host: string;
   protocol: string; // not the same as the Protocol for binding
   protocolVersion?: string;
-  security?: SecurityRequirementObject[];
-  bindings?: Bindings<SocketIOServerBinding>;
+  pathname?: string;
+  title?: string;
+  description?: string;
+  variables?: Record<string, ServerVariableObject>;
+  security?: SecuritySchemeObject[];
+  tags?: TagObject[];
+  externalDocs?: ExternalDocumentationObject;
+  bindings?: Bindings<WSServerBinding>;
 }
 
+/** @since 3.0.0 contains tags and externalDocs */
+export interface InfoObject extends OASInfoObject {
+  tags?: TagObject[];
+  externalDocs?: ExternalDocumentationObject;
+}
+
+/** @since 3.0.0 channels are optional, added operations */
 export interface AsyncApiObject {
   asyncapi: string;
   /** @desc URI or URN format */
   id?: string;
-  info: OASInfoObject;
+  info: InfoObject;
   servers?: Record<string, ServerObject>;
-  channels: ChannelsObject;
+  channels?: ChannelsObject;
+  operations?: OperationsObject;
   components?: ComponentsObject;
-  tags?: TagObject[];
-  externalDocs?: ExternalDocumentationObject;
   defaultContentType?: string;
 }
 
 /**
- * @desc A relative path to an individual channel. The field name MUST be in the form of a RFC 6570 URI template.
- * @desc Query parameters and fragments SHALL NOT be used, instead use bindings to define them.
+ * @since 3.0.0 An identifier for the described channel. The channelId value is case-sensitive.
  * */
-export type ChannelsObject = Record<string, ChannelItemObject>;
+export type ChannelsObject = Record<string, ChannelObject>;
 
-export interface ChannelItemObject {
+/** @since 3.0.0 renamed; added address, title, summary, messages, servers, tags, externalDocs; removed pubs/subs */
+export interface ChannelObject {
+  /** @desc Typically the "topic name", "routing key", "event type", or "path". */
+  address?: string | null;
+  title?: string;
+  summary?: string;
   description?: string;
-  /** @desc the messages produced by the application and sent to the channel. */
-  subscribe?: OperationObject;
-  /** @desc the messages consumed by the application from the channel. */
-  publish?: OperationObject;
+  messages?: MessagesObject;
+  servers?: ReferenceObject[];
   /** @desc Describes a map of parameters included in a channel name. */
   parameters?: ParametersObject;
+  tags?: TagObject[];
+  externalDocs?: ExternalDocumentationObject;
   /** @desc Map describing protocol-specific definitions for a channel. */
-  bindings?: Bindings<SocketIOChannelBinding>;
+  bindings?: Bindings<WSChannelBinding>;
 }
 
 export interface ServerVariableObject extends OASServerVariableObject {
   examples?: string[];
 }
 
-/**
- * @desc Each name MUST correspond to a security scheme which is declared in the Security Schemes under the Components.
- * @desc If the security scheme is of type "oauth2" or "openIdConnect", then the value is a list of scope names.
- * */
-export type SecurityRequirementObject = Record<string, string[]>;
-
+/** @since 3.0.0 added replies */
 export interface ComponentsObject {
   schemas?: Record<string, SchemaObject>;
   servers?: Record<string, ServerObject>;
   serverVariables?: Record<string, ServerVariableObject>;
-  channels?: Record<string, ChannelItemObject>;
+  channels?: Record<string, ChannelObject>;
   messages?: Record<string, MessageObject>;
   securitySchemes?: Record<string, SecuritySchemeObject>;
   parameters?: Record<string, ParameterObject>;
   correlationIds?: Record<string, CorrelationIDObject>;
   operationTraits?: Record<string, OperationTraitObject>;
   messageTraits?: Record<string, MessageTraitObject>;
-  serverBindings?: Bindings<SocketIOServerBinding>;
-  channelBindings?: Bindings<SocketIOChannelBinding>;
-  operationBindings?: Bindings<SocketIOOperationBinding>;
-  messageBindings?: Bindings<SocketIOMessageBinding>;
+  replies?: Record<string, OperationReplyObject>;
+  serverBindings?: Bindings<WSServerBinding>;
+  channelBindings?: Bindings<WSChannelBinding>;
+  operationBindings?: Bindings<WSOperationBinding>;
+  messageBindings?: Bindings<WSMessageBinding>;
 }
 
+/** @since 3.0.0 supports MultiFormatSchemaObject in payload */
 export interface MessageObject extends MessageTraitObject {
-  payload?: SchemaObject | ReferenceObject;
+  payload?: SchemaObject | MultiFormatSchemaObject | ReferenceObject;
   traits?: MessageTraitObject | ReferenceObject;
 }
 
+/** @desc The key represents the message identifier. The messageId value is case-sensitive. */
+export type MessagesObject = Record<string, MessageObject | ReferenceObject>;
+
+/** @since 3.0.0 added action, channel, messages, reply */
 export interface OperationObject extends OperationTraitObject {
+  action: "send" | "receive";
+  /** @desc A $ref pointer to the definition of the channel in which this operation is performed. */
+  channel: ReferenceObject;
+  /** @desc A list of $ref pointers pointing to the supported Message Objects that can be processed by this operation */
+  messages?: ReferenceObject[];
+  reply?: OperationReplyObject | ReferenceObject;
   traits?: Record<string, OperationTraitObject>;
-  message?:
-    | {
-        oneOf: Array<MessageObject | ReferenceObject>;
-      }
-    | MessageObject
-    | ReferenceObject;
 }
 
+/**
+ * @desc Describes the reply part that MAY be applied to an Operation Object.
+ * @desc If an operation implements the request/reply pattern, the reply object represents the response message.
+ * @since 3.0.0 new
+ * */
+interface OperationReplyObject {
+  /** @desc Definition of the address that implementations MUST use for the reply. */
+  address?: OperationReplyAddressObject | ReferenceObject;
+  /** @desc A $ref pointer to the definition of the channel in which this operation is performed. */
+  channel?: ReferenceObject;
+  /** @desc A list of pointers to the supported Message Objects that can be processed by this operation as reply */
+  messages?: ReferenceObject[];
+}
+
+/**
+ * @desc An object that specifies where an operation has to send the reply.
+ * @since 3.0.0 new
+ * */
+interface OperationReplyAddressObject {
+  /**
+   * @desc A runtime expression that specifies the location of the reply address.
+   * @example $message.header#/replyTo
+   * @example $message.payload#/messageId
+   * @link https://www.asyncapi.com/docs/reference/specification/v3.0.0#runtimeExpression
+   * */
+  location: string;
+  description?: string;
+}
+
+/**
+ * @desc The operation this application MUST implement. The field name (operationId) MUST be a string used to
+ * @desc identify the operation in the document where it is defined, and its value is case-sensitive.
+ * */
+export type OperationsObject = Record<string, OperationObject>;
+
+/** @since 3.0.0 operationId moved to OperationsObject; added title, security */
 export interface OperationTraitObject {
-  /** @desc Unique string used to identify the operation. */
-  operationId?: string;
+  title?: string;
   summary?: string;
   description?: string;
+  security?: SecuritySchemeObject[];
   tags?: TagObject[];
   externalDocs?: ExternalDocumentationObject;
-  bindings?: Bindings<SocketIOOperationBinding>;
+  bindings?: Bindings<WSOperationBinding>;
 }
 
+/** @since 3.0.0 messageId moved to MessagesObject, schemaFormat moved to MultiFormatSchemaObject */
 export interface MessageTraitObject {
-  /** @desc Unique string used to identify the message. The id MUST be unique among all messages described in the API.*/
-  messageId?: string;
-  headers?: SchemaObject;
+  headers?: SchemaObject | MultiFormatSchemaObject;
   correlationId?: CorrelationIDObject;
-  schemaFormat?: string;
   contentType?: string;
   name?: string;
   title?: string;
@@ -125,7 +176,31 @@ export interface MessageTraitObject {
   description?: string;
   tags?: TagObject[];
   externalDocs?: ExternalDocumentationObject;
-  bindings?: Bindings<SocketIOMessageBinding>;
+  bindings?: Bindings<WSMessageBinding>;
+  examples?: MessageExampleObject[];
+}
+
+/**
+ * @desc Represents an example of a Message Object and MUST contain either headers and/or payload fields.
+ * @since 3.0.0 new
+ * */
+interface MessageExampleObject {
+  name?: string;
+  summary?: string;
+  headers?: Record<string, any>;
+  payload?: Record<string, any>;
+}
+
+/** @since 3.0.0 new */
+interface MultiFormatSchemaObject {
+  /**
+   * @desc A string containing the name of the schema format that is used to define the information.
+   * @example application/vnd.aai.asyncapi+yaml;version=3.0.0
+   * @example application/schema+yaml;version=draft-07
+   * @example application/vnd.oai.openapi+yaml;version=3.0.0
+   * */
+  schemaFormat: string;
+  schema: SchemaObject;
 }
 
 export interface CorrelationIDObject {
@@ -133,6 +208,7 @@ export interface CorrelationIDObject {
   /**
    * @desc A runtime expression that specifies the location of the correlation ID.
    * @example $message.header#/correlationId
+   * @link https://www.asyncapi.com/docs/reference/specification/v3.0.0#runtimeExpression
    * */
   location: string;
 }
@@ -185,10 +261,13 @@ export interface OAuthFlowObject {
   scopes: Record<string, string>;
 }
 
-export interface ParameterObject {
-  description?: string;
-  schema?: SchemaObject | ReferenceObject;
-  /** @desc A runtime expression that specifies the location of the parameter value. */
+/** @since 3.0.0 partially extends SchemaObject; schema prop removed */
+export interface ParameterObject
+  extends Pick<SchemaObject, "enum" | "default" | "description" | "examples"> {
+  /**
+   * @desc A runtime expression that specifies the location of the parameter value.
+   * @link https://www.asyncapi.com/docs/reference/specification/v3.0.0#runtimeExpression
+   * */
   location?: string;
 }