+++ src/async-api/ws-binding.ts
@@ -1,16 +1,17 @@
 import type { ReferenceObject, SchemaObject } from "openapi3-ts/oas31";
-
-// @todo use 1.0.0 when production ready
-export const version = "0.11.0";
-
 /**
- * @desc Protocol-specific definitions for a server.
+ * @see https://github.com/asyncapi/bindings/tree/master/websockets
  * @desc This object MUST NOT contain any properties. Its name is reserved for future use.
  */
-export interface SocketIOServerBinding {}
+export interface WSServerBinding {}
 
-/** @desc Protocol-specific definitions for a channel. */
-export interface SocketIOChannelBinding {
+/**
+ * @desc When using WebSockets, the channel represents the connection. Unlike other protocols that support multiple
+ * @desc virtual channels (topics, routing keys, etc.) per connection, WebSockets doesn't support virtual channels or,
+ * @desc put it another way, there's only one channel and its characteristics are strongly related to the protocol
+ * @desc used for the handshake, i.e., HTTP.
+ */
+export interface WSChannelBinding {
   /**
    * @desc The HTTP method to use when establishing the connection. Its value MUST be either GET or POST.
    * */
@@ -28,22 +29,13 @@ export interface SocketIOChannelBinding {
   /**
    * @desc The version of this binding. If omitted, "latest" MUST be assumed.
    */
-  bindingVersion: typeof version;
+  bindingVersion: "0.1.0";
 }
 
 /**
- * @see https://github.com/asyncapi/bindings/tree/master/websockets
- * @desc Protocol-specific definitions for an operation.
- * @desc This object MUST NOT contain any properties. Its name is reserved for future use.
- */
-export interface SocketIOOperationBinding {}
-
-/**
- * @desc Protocol-specific definitions for a message.
  * @desc This object MUST NOT contain any properties. Its name is reserved for future use.
  * */
-export interface SocketIOMessageBinding {
-  /** @desc Acknowledgement schema */
-  ack?: SchemaObject | ReferenceObject;
-  bindingVersion: typeof version;
-}
+export interface WSOperationBinding {}
+
+/** @desc This object MUST NOT contain any properties. Its name is reserved for future use. */
+export interface WSMessageBinding {}